Updated some of the gconf key locations, now sounds are not in a "sounds"
* gossip.schemas.in: Updated some of the gconf key locations, now
sounds are not in a "sounds" parent of their own, they are in a
"notifications" parent since we now have popup preferences too.

* ui/main.glade:
* src/gossip-notify.c:
* src/gossip-preferences.[ch]: Reworked the preferences a bit so
we have the gconf keys defined in ONE place instead of everywhere
and added a preference to be able to disable the "Joe Bloggs has
become	available" notification.

* src/gossip-app.c:
* src/gossip-chat.c:
* src/gossip-sound.c:
* src/gossip-status-presets-dialog.h:
* src/gossip-theme-manager.c:
* src/gossip-vcard-dialog.c: Use the new defined gconf keys in
gossip-preferences.h.