Remove old use of gossip_sound_set_enabled
We used to connect to the contact list to play a sound when a user was added/removed to it. It meant that we played sounds for every offline contact when we switched to "Show offline contacts". However, after the libgossip refactoring we now connect to GossipSession instead and play the sound based on the contacts presence, rather than whether it's shown in the contact list or not.

This code currently do nothing so remove it.

svn path=/trunk/; revision=2328