Remove status icon and mac dock support.
Status icon has gone out of fashion in GNOME and the API is deprecated
in GTK+. Remove them from the code base and disable the toggling of
main window.

The main window will be changed to contain both the roster and chat
window to fit better into GNOME 3.0 where single app window seems to
be the way to go.